moved summarize sections into extract document structure
