+++ lib/sycamore/sycamore/transforms/extract_document_structure.py
@@ -114,3 +114,29 @@ class ExtractDocumentStructure(Map):
         **resource_args,
     ):
         super().__init__(child, f=structure.extract, **resource_args)
+
+
+class ExtractSummaries(Map):
+    """
+    Extracts summaries from child documents to be used for entity extraction. This function
+    generates summaries for sections within documents which are used during entity extraction.
+    """
+
+    def __init__(self, child: Node, **resource_args):
+        super().__init__(child, f=ExtractSummaries.summarize_sections, **resource_args)
+
+    @staticmethod
+    def summarize_sections(doc: HierarchicalDocument) -> HierarchicalDocument:
+        if "EXTRACTED_NODES" in doc.data:
+            return doc
+        for section in doc.children:
+            assert section.text_representation is not None
+            summary = f"-----SECTION TITLE: {section.text_representation.strip()}-----\n"
+            for element in section.children:
+                if element.type == "table":
+                    element.text_representation = element.data["table"].to_csv()
+                assert element.type is not None
+                assert element.text_representation is not None
+                summary += f"""---Element Type: {element.type.strip()}---\n{element.text_representation.strip()}\n"""
+            section.data["summary"] = summary
+        return doc
